Fix revert of placeholder.

+++ s3/src/main/java/ch/cyberduck/core/s3/S3VersioningFeature.java
@@ -187,11 +187,12 @@ public class S3VersioningFeature implements Versioning {
                 destination.setAcl(accessControlListFeature.convert(acl));
                 session.getClient().copyVersionedObject(file.attributes().getVersionId(),
                     containerService.getContainer(file).getName(), containerService.getKey(file), containerService.getContainer(file).getName(), destination, false);
-                if(file.getParent().attributes().isDuplicate()) {
+                if(file.getParent().attributes().isDuplicate() &&
+                    file.getParent().attributes().getVersionId() != null) {
                     // revert placeholder
                     session.getClient().deleteVersionedObject(
                         file.attributes().getVersionId(),
-                        containerService.getContainer(file).getName(), containerService.getKey(file));
+                        containerService.getContainer(file).getName(), containerService.getKey(file.getParent()));
                 }
             }
             catch(ServiceException e) {
